workbench-summarizer/requirements.txt
google-cloud-documentai
google-cloud-storage
google-cloud-bigquery
functions_framework